# Optimize array construction and allocation