For 8 bytes, from 1.8 us to around 1 us, closer to the 800ns from before. For reference, before the stream-ordered allocator an 8-byte allocation only took around 500ns when it hit the pool. Base allocates in 20ns or so...